Tasks/Notes/Reporting Finished
CRUD of tasks, and comments should be completely completed as of this
commit. Additionally, made some changes to the majority of the links
present to ensure that there was no errors with swapping pages. There's
probably more fixed here...can't remember what though, worked on it most
of the day, non-stop - no access to Git though.